Refactor bind database in Binder into an ICommand.
In an effort to massively improve the code in the Binder, the gigantic
class is being broken into ICommand based classes. This commit refactors
the database binding code by simply extracting the existing code from
the Binder into the BindDatabaseCommand and friends. Improving the code
comes after this fundamental refactoring step.