WIXBUG:4114 - PostQuitMessage in wixstdba from WM_NCDESTROY.
SHAutoComplete documentation says you need to wait for the control to be
destroyed before calling CoUninitialize. Since the window gets the
WM_DESTROY message before its child controls, the control is probably
never being destroyed. If we wait for the WM_NCDESTROY message instead,
the control will be destroyed when we call PostQuitMessage.