Remove non-Field backed properties from File rows using FileFacade.
Major refactor to FileRow to remove properties that were set during
binding but not backed by Fields and thus not persisted to the .wixpdb.
That is not a desirable pattern so remove them and fix the code. The
pure WiX patching solution is broken at this point and needs to be
fixed with a future check-in.