Improve anonymous id generation for Components
Anonymous Components get their identity from their key path. Unfortunately,
when determining the key path, the identity of the Component is often
required. This chicken and egg problem is solved by temporarily using a
binder variable as the Component Id then after the key path is calcuated
immediately resolving Component Id binder variable in all rows in the
active section.

Immediately resolving the anonymous Component's binder variable also means
that as the active section is repeatedly scanned as each anonymous
Component is added. For very large sections with many anonymous Components
this can create a serious performance problem.

The fix is to save all anonymous Component binder variable resolution
until the end of the compile for the entire intermediate. Doing so means
the compiled sections are scanned a single time. This is extremely
efficient.