Minimize creation of simple references and anonymous directories
While it is legal to duplicate simple references and anonymous
directories in a section it is wasteful. The intermediate file is larger
and creates redundant work for the linker. This change tracks simple
references and anonymous directories and avoids creating duplicates which
can significicantly reduce the duplication found in intermediates.